third_party/blink/web_tests/external/wpt/web-animations/interfaces/Animation/commitStyles-expected.txt

This is a testharness.js-based test.
[FAIL] Commits logical properties
  assert_equals: expected "20px" but got "10px"
[FAIL] Commits logical properties as physical properties
  assert_equals: expected "20px" but got "10px"
Harness: the test ran to completion.
